fix(movement): stop arm64 fusing the heading's products into the wrong answer
The determinism matrix found it on its first real run. linux-arm64,
macos-arm64, and windows-arm64 disagreed with both amd64 targets on exactly
three of the five recordings — ice, slime-and-soul-sand, and sprint-diagonal —
and agreed on the other two. The three are precisely the runs where strafe and
forward are both non-zero, which is where the heading has two products feeding
an add.

Go permits an implementation to contract a multiply and an add into one
operation with a single rounding, unless an explicit conversion rounds the
intermediate. On arm64 the compiler takes that permission. The fix is the
conversion the spec names: each product is narrowed to float32 before the
addition, in the heading and in the cosine's table index, where the quarter turn
is added to a product for the same reason.

This is a correctness bug, not a tie between two defensible answers. Java never
contracts unless Math.fma is called, so the game computes each product
separately, and internal/oracle checked this module's unfused arithmetic against
a real 1.8.9 server over 4,800 ticks. An arm64 build was producing positions
vanilla does not produce.

The committed recordings are unchanged and still verify here, which is the
evidence that the conversions force the answer amd64 already had rather than
choosing a third one. That is also why no recording was regenerated: the plan
forbids it as a response to a red matrix, and it was not needed.

+++ movement/heading.go
@@ -49,7 +49,7 @@ func radians(yaw float32) float32 { return yaw * pi / 180 }
 // Widening earlier than step four changes the result in its last bits, and a
 // hundred ticks of that is a position a server corrects.
 func ApplyHeading(table Table, motion geom.Vec3, strafe, forward, speed, yaw float32) geom.Vec3 {
-	magnitude := strafe*strafe + forward*forward
+	magnitude := float32(strafe*strafe) + float32(forward*forward)
 	if magnitude < inputThreshold {
 		return motion
 	}
@@ -67,8 +67,25 @@ func ApplyHeading(table Table, motion geom.Vec3, strafe, forward, speed, yaw flo
 	sin := table.Sin(angle)
 	cos := table.Cos(angle)
 
-	motion.X += float64(strafe*cos - forward*sin)
-	motion.Z += float64(forward*cos + strafe*sin)
+	// Every product is converted to float32 before it is added or subtracted,
+	// and the conversions are load-bearing rather than decorative.
+	//
+	// Go permits an implementation to fuse a multiply and an add into one
+	// operation with a single rounding, unless an explicit conversion rounds the
+	// intermediate. On arm64 it takes that permission: the determinism matrix
+	// found linux-arm64, macos-arm64, and windows-arm64 disagreeing with both
+	// amd64 targets on exactly the three recordings where strafe and forward are
+	// both non-zero — the runs where these expressions have two products to fuse.
+	// The runs with forward alone agreed, because the other product is zero and
+	// fusing it changes nothing.
+	//
+	// The unfused answer is the correct one, not merely the majority one: Java
+	// never contracts a multiply and an add unless Math.fma is called, so the
+	// game computes each product separately, and the oracle checked this
+	// module's unfused arithmetic against the game over 4,800 ticks. An arm64
+	// build without these conversions was producing positions vanilla does not.
+	motion.X += float64(float32(strafe*cos) - float32(forward*sin))
+	motion.Z += float64(float32(forward*cos) + float32(strafe*sin))
 
 	return motion
 }
